Rettet id til Long

+++ src/main/java/com/padc/demo/user/controller/UserController.java
@@ -65,7 +65,7 @@ public class UserController {
     }
 
     @GetMapping("/bruger/bruger_side/{id}")
-    public String showUser(@PathVariable("id") int id, Model model) {
+    public String showUser(@PathVariable("id") Long id, Model model) {
         try
         {
             User user = userService.findById(id);
